Implement base recolouring infrastructure
we manipulate the ramp texture selection curve to change the colour of
the factions

so far this covers: all terrain and idle effects, specified in the unit
effect blueprint.

an example here is provided: the idle hover effect on the nomads tank is
recoloured - note that its only on land and only idle, but takes only 1
more line to affect water as well.

an example ramp and psd file is included, it can be transferred from
ramp to ramp but its important to ensure that the ramp is 32px high and
that the banding on it is exactly like that for the bottom 20 bands. the
rest are empty space for later use perhaps?