Batch retexture all albedo files
this is the first step in the nomads retexturing.
the purpose is to deliver maximum improvement in the least amount of
time. a proper retexture would take a lot more time but this gets us to
a starting point very quickly.

this needs a bunch of tweaks, or if possible get brutes texture tool to
work.