Rename nomad to Nomads in a most absurd commit
holy shit.

bscly, there was a lot of nomads and a lot of nomad and there was no
pattern. this changes as much as possible into nomads and so use that
for everything in the future.

HOOOOOOWEVER:
because of literal black magic i could not get the shaders to be
affected, SO, things related to shaders are still Nomad. but its now a
very small part of the code.

merge if you dare.