building effects to inc0001
self:AddToConstructionQueue('inu1001', 'Player1')

will create a t1 engy for the player 'Player1'.
The loop that is in the OnCreate function will create an engy for each
of the players in the mission that are not hostile with the ship (i.e.
all humans)

This function can be used to create any unit for any player.