Fix for "No Transport Full Sound"
The T1 Nomad Transport was importing the `NAirUnit` class from `nomadsunits` to its script instead of `NAirTransportUnit`. I would imagine the reason you would have to spawn another transport to get it to work is because the transport class isn't imported by the T1 Nomad Transport script when it's created. However, the class is imported when any other transport is created. Changing the import for the T1 Nomad Transport has fixed all audio errors without affecting the transport behaviour.  Tested and ready for verification. Enjoy. 

P.S. For future reference. The "Transport Full" announcement is handled in the `aibrain.lua` file.