Replace Nomads T1 Tank Destroyer
yeah, pretty much this. new model, textures, animations!
the balance is pretty much the same as it was before though.
the anchor ability is back but it doesnt add any benefits, its a purely
immobilisation option for now. this means its not really needed, mostly
to preview the animation and how that works.

while this works, it could be rebalanced to make it even better.