Refactor Nuke code and explosions
yeah so this is a huge amount of stuff changed, everything to do with
nukes.

1.double explosion damage fixed
2.veterancy dispersal issues fixed
3.better compatibility for future changes
4.code cleaned up, cleaner logic

what happens now is that the damage is done through the projectiles like
normal. The projectiles create an entity which does all sorts of fancy
effects. units killed by a special damage type, are sucked in by the
entity.
the parent unit should have a self.NukeEntity value so that if it dies
the entity can still function.