Integrate support for campaign missions
previously this was in a separate mod, and to add extra flexibility and
support, we integrate it so that you can run nomads missions with one
less mod!

note, you still need the coop mod enabled for them to work, and some
changes that are not appropriate to base nomads (balance adjustments for
example) wouldnt be present.