Fix issue with T2 units loading into T2 transport
so, when you order t2 units into an empty transport, they fail to all
fit, with the transport only letting one on. This happens because it is
folded up and unfolds. When unfolded then you can load the rest.

i suspect this is due to the attachment bones being used in the
animation, since t1 units dont have this issue.

this has been fixed by making the transport unfold once when built and
stay in that state, like other transports.