Fix ACU OC functionally
this makes the overcharge not broken. now it drains energy correctly and
doesnt cause errors in the log.

because the equation for OC damage is non-linear, its insane to attempt
to split it in between projectiles in a burst, so the overcharge has
been changed to a single projectile.

there still needs to be a cosmetic distinction between regular
projectiles and OC, which needs to be done at some time later.