Refactor Capacitor
now the code is a lot cleaner and easier to deal with!
the capacitor exists in 4 states and cycles between them. the state
variable is used for all the state specific logic on both ui and sim
side, so its easy to track.

When unfilled, you can start charging the capacitor, which drains
energy. You can cancel halfway through, and the capacitor will decay to
empty. If the capacitor is charged fully, it stops draining energy and
can be activated to provide buffs for a limited time, after which it
becomes empty.

and thats it!

more things coming soon, such as proper recognition of mixed selections,
better drain mechanisms, and possible better balance!