Rewrite Damage to shields
the nomads units were doing it in a different way, and now they do it in
the same way as FAF, by passing data in the event that they hit a
shield, with added overkill controls.

this requires script support for every unit individually, but this is
likely a good thing since it has a performance boost for all projectiles
that dont need it.

it would have been nice to have a unified data passing function in the
base code but one can only hope xD