Rewrite Orbital Unit support
So, now frigates are called by units instead of globally and respond to
units instead of globally as well. Overall this is much more error
resistant and less likely to cause crashes, and more flexible, too!

As part of this commit, the way artillery satellites are spawned has
been redone as well. Now they spawn from the nearest frigate up to 500
units away and if its too far or the frigate is too busy, additional
frigates are summoned to help out, and then leave afterwards.