Remove Tasks.lua from hooks
This file isnt used by FAF at all. No idea why its there at all, which
means that its better to not hook it for nomads, making it independent
of the FA repo.

This is part of the abilities code rewrite. Perhaps this file will be
removed entirely in the future.