Remove customisation of ability methods
the issue with this was that it was not used, and added complexity to
the code, in addition to importing files mid function calls often, which
is a bit of a performance drain.

If this becomes needed in the future for whatever reason, it could be
added in later. possibly in a different way.
Lastly, there was one import left which is kept just in case it IS,
used, which should not happen. It will be removed at a later date if
theres no evidence of use.