Reduce build beam lag
this is sort of a quick fix - the beams were cycled through too often
and so were causing a load of lag. this was fixed by making the cycles
rarer.

in the future if its possible the beams should be rewritten so that each
cycle is also less load individually.