Rewrite Orbital frigate movement mechanics
now the frigate has two modes: follow, and stay still.

the acu (and other units) can order it to follow a unit, stay still, or
move to a specific location.

when the acu has a relevant upgrade, it follows that acu. when it
doesnt, it stays still.

this is more robust than before, and if multiple acus give it commands
to follow, it will complete the last one.