Update Orbital bombardment to use projectiles
now the ability relies on needing to have counted tactical projectiles
stored in the unit to initialise orbital bombardment.

the cooldown on the ability is also removed, since the rate of fire is
determined by the projectiles being built.

This also supports multiple ACUS, but in a wonky way - they all trigger
the same launch sequence - the frigate cant support queing them. Either
one frigate per ACU, or selecting the ACUS one at a time and launching
is needed to not waste multiple projectiles per launch right now. needs
to be fixed later.