Add shader support for recoloured emitters
pretty much what it says on the tin. This is some of the most awful and
terrifying code i have ever written.

We cant add more variables to the shaders, since they are governed by
the engine, so the new shaders use existing variables to sneak the
values in, and then change the hue of the emitters.

Well, it works at least.