Reorganise Weapons and Projectiles
This is the start of a very long refactor indeed :>

The old weapons and projectiles arent very clearly organised, and so
they need to be combined, merged, rearranged and so on.

To help this, there will be two lists - the legacy list and the new
list. As projectiles are updated, they will be moved from the legacy
list, and gradually we will end up with a cleaner set of files.