Rework SCU upgrades
this is an absolutely huge rewrite which takes a look at all of its
upgrades and pretty much does them again from scratch.

1.there are no longer any armless varieties
2.the upgrades have been thoroughly reshuffled - some removed, some
moved slots, renamed, and some added

now there are a couple of unique options for each slot, with no options
to duplicate upgrades like before. Overall the number of upgrades is
greatly reduced.

A whole lot of script cleanup has happened as well.