Fix Artillery creating multiple firing sounds
this has no effect on gameplay, but it fixes the audio of the artillery
unit. For some reason, weapons call OnFire and play the sound for Fire
there, and also in CreateProjectileAtMuzzle. Nomads satellites call this
manually so it creates the sound from multiple places. This is fixed by
changing the sound name and playing it manually.