Fix issue with global variables for abilities
this was causing a problem where global variables were being used
instead of local ones, making the tables synced when they shouldnt be.
specifically, when you changed one ability on one unit, it would update
all other units with that ability, so all unit either had it or didnt.
This has been fixed with each entry being unique now.

Also, there was a bunch of code cleanup, with legacy functions being
removed and the like.