Allow abilities to update mid activation
among other cleaning, this allows units killed or built or activated to
update the abilities while they are being used, so you dont have to
cancel and order them again. Also there is a bunch of other cleaning, as
well as fixing a bug where reticule positions were not being set
correctly.