Fix beamer weapon packing issue
When the beamer was packing, and needed to unpack again, the animation
was not overidden because it has a continuous beam and thats set a bit
lousy in the default weapons. turning off the flag whenever the beam is
actually disabled solves the issue and stops the beamer from getting
stuck mid animation.