Update Command frigate animations
Now there is a takeoff animation, which can be used whenever the frigate
needs to warp out. Right now its not used anywhere but the support
exists. This is planned for the campaign, as well as for removing
temporary frigates for spawning satellites. Additionally there might be
a use for them to fly off if their player is defeated.