Refactor Command frigate animation sequences
this is just some cleaning up, and rearranging the code a little so
various functions could be called later if needed by things like
missions and other scripts, allowing the frigate to take off, to reset
its rotators to a rotating state, and so on. Also fixes a minor bug with
the warp in animation, where the warp armour bone was retracted a bit
too far.