Refactor updateBuildRestrictions in unit.lua
this was hooked because the original function hardcoded factions. also,
the original function is insane. now its a lot more sane, and even
supports dynamic faction lists. at some point this should go into the fa
repo, but theres no need to wait before putting it in here.

also, it should be less laggy since it loops through lists only when it
needs to now instead of every time for everything :>