Rework T1 MAA/A Targeting behaviour
previously the maaa would target air too much or when toggled be a pain
in the ass due to being unable to target ground units, which made the
unit pretty frustrating to use.

Now its behaviour is much smarter, allowing manual target overrides and
similar. Overall this should need much less toggling and be much nicer
to use.