Refactor Capacitor to be more robust when removed
Pretty much this, there were a couple of cases where the capacitor could
be removed and do weird things, such as continue charging. This was
mostly a cosmetic issue, but now it should be completely robust.
Additionally some one line function stuff has been removed (sacu file is
gonna be left for later, that needs a massive rewrite anyway)