Refactor Intel Boost as a toggle in NomadsUnits
now the code is shared between all units. Requires the commits cleaning
up the unit scripts.

The toggle is much simpler, and this is futureproofed, with support for
things like animations and magic like that when its added. Currently the
effects are changed when the radar is toggled on, and when the unit dies
while boosted it also EMPs stuff around it.