Refactor counter Ammo ability code
So, the code was a bit messy, and now its nicer. theres no need for
different functions for adding and removing ammo count, since the end
result is the same. The SetSpecialAbilityAvailability now is a check for
sanity, and toggles one ability at a time as opposed to all of them.

Also the original factions acus have their hooked functions changed so
they dont manipulate the blueprint anymore which was a messy way of
doing things and instead just update the code directly, since thats now
supported.