Refactor Meteor Script to better deal with trees
This is a partial fix that clears trees in the starting area. For some
reason there can be cases where some are left over, and it still needs
to be worked out why, but in any case the situation is much improved.
Also the code is cleaner too.

The final fix may have to wait until the warp in animation is redone.