Add warning to T4 beamer in case of packing bug
This is mostly a just in case measure, to try and find a rare bug, which
seems to be difficult to reproduce. The bug is where the beamer fires
when its weapon isnt unpacked. Doesnt change much functionally but
should trigger if the edge cases happen, even though they shouldnt.