Refactor Transport setting code in blueprints.lua
Pretty much what it says on the tin. T4 transports need to be able to
move naval units, and this code sets their blueprints to allow that. The
functionality is the same as before except now we use the hashed
categories table, and dont do a bunch of useless stuff. So the code is 3
times shorter.