Update T4 transport model
Now the model contains better bone hierarchy for its engines, its Sams
are in a nicer default position, and it has exhaust bones for its rear
engines so that effects can be added there later. Also, there were a
couple of vertices that werent split and were causing graphical errors.