Fix Capacitor duration issue
This is actually a pretty big bug, its due to the IsStunned variable,
which happens to be the exact same thing as the engine function. It was
renamed from _IsStunned to make things more clear, but instead we use
the engine function since thats what its there for. the variable being
there at all is rather bizarre.