Adjust Orbital bombardment missile movement
Now the random spread takes terrain height into account, while
previously it assumed that the terrain was flat, which led to issues
when aiming at mountains.

Additionally the wait times for the staging have been changed a little
so the missiles check more often on the earlier stages, where they are
moving super fast. This avoids the salvos spreading out during travel as
much.