Rewrite Beamer death sequence
Mostly this is to add support for it being able to sink, but it also
cleans up a bunch of code, and makes the beam disable correctly if it
was turned on when the unit died.

This could still do with some improving, right now its mostly a
functional change, but it could be perhaps made more beautiful as well.