Refactor blackhole callbacks in proptree.lua
This is a small cleanup, this function isnt needed to be called anywhere
else so we can just put it right into the  behaviour.

This also removes issues with the black hole code not working, since it
was being called wrong:
self:DoBlackHoleCallbacks(self, instigator) is not how you do it.