Fix vanilla units stats showing wrong
Now these units show the correct units stats due to black magic in their
scripts. We also fix this because we can and because people keep trying
to look up stats in nomads to compare with nomads and expect it to
actually work. so we need to fix faf bugs too.