Adjust Cruise missile projectile for MML
The turn rate on this was too low at the longest range of the MML so now
its a bit longer, which makes the mml behave nicely, as well as the TML
which uses the same projectile.

Possibly, it could be better to use different ones for the different
units.