Enable Backwards Compatibility with FAF Refactor
So, the refactoring from speed is lovely but it relies on his other work
in the fa branch, so this just adds the functionality into nomads
directly as well, so in case the fa work disappears nomads will still
work. At some point in the future this will be removed once it makes it
to released status, but for now its good to be independant and be able
to release stuff without explosions.