Add nomads notify messages
This adds notify messages for nomads. an issue is that they way notify
works, this is mostly useless right now since you need to reset your
notify from the menu to get the new messages. It would be good to check
if they are there and then add them in if theyre missing on startup.