Fix Missile spread function not working on sea
The missiles that were retargetted by this function ended up shooting
below the target, and so hitting the water and missing.

So the issue here was that the terrain height got the surface of the
seabed, while the point was to get the surface of the water. Luckily
Surfaceheight does just that, so it now retargets missiles on the
surface of the water, meaning they wont miss anymore.