Fix Intel probe created many times per tick
This is in theory an abuse - what you do is you queue an order to the
acu and then spam the intel probe command. while its walking you get to
order as many as you like and then it all spawns them in the same tick
so the system is confused whether it created one or not. This prevents
multiple probes appearing in the same tick.