Fix Lav Root Bone Name
Previously it was the old INU1007 and now its the new name, this should
have been named right from the start but oh well, fixed now.

Also this is testing out a new version of the exporter which features
slightly smaller model file sizes due to added efficiency so thats
great.