Let ACU and Exps use the new snipe mode
So one thing i missed was that snipe mode has a new limiter on it thats
supposed to prevent units from sniping, for whatever reason.

This means that it doesnt work if you dont have the special category, so
adding the category fixes that. whatever i guess. the whole
implementation of the limited is fairly weird since in most cases what
happens is that snipe mode actually decreases ACU priority over what is
the default, so its rather buggy.

Not a nomads bug though so fixing it here isnt really in the scope of
the mod.