Rename Rapid Repair variables
Apofenas messed it up a little, the upgrades stack per upgrade slot and
replace each other, that way you need only 4 variables in total - 1 per
slot and one base.

Anyway renamed that as well as making the names the same for all
upgrades in the blueprint so they arent super confusing.