Update T4 transport to have animated death
Pretty much this. Thanks to epic work by Clockwork, the transport now
has some fancy animations for its engines exploding. So we apply those
animations randomly to one side at a time!

This does need some polish however, both with one of the animations
needing some adjustment as well as a little bit of bugfixing along the
way, such as having an issue with the transport hitting the water.