Reduce Beamer getting targets stuck
So, after a long time of pulling out hair, it turns out this is a weird
bug to do with aim controllers - a unit that has a large tracking radius
can get things stuck inside. This happens when a unit has something just
outside of its range thats of a higher priority than anything inside.

Not a complete solution, but this is an improvement, and also it will
help the beamer be more convenient in general since its able to pack and
unpack faster