Futureproof ingame unit stats
FAF is going to destructively replace this function, and its totally
cursed. This is the only way i could think of to make it work - use a
fixed version of the current code with a switch that detects arguments,
and then use the switched version of the future code.

Amazingly bad code but it will get mostly cleaned up once faf releases
their patch.