Fix T4 transport build ui breaking
So the function that defines how much a transport can transport is badly
coded since it can just return nil values in some cases which breaks
everything. Now it cant, and also now it returns the t4 transport value.

In theory you want to count up the number of bones in the transport with
the right name and then use that to work out the number but thats an
excercise left to the reader. As if anyone will ever read this.