Remove error spam on ability panel initialisation - partial fix
This doesnt fully fix the ability panel ui but it does make it not spam errors. The issue was trying to assign table values to the NewButtonGlows when it was a bool by default - not sure why that was written like that.

This also means that something else that set it to a table stopped working?